requirements.txt
pypdf2
BeautifulSoup4
spacy
langchain
python-dotenv
predictionguard
langchain-together
monsterapi==1.0.2b3